+++ src/constants/constants.js
@@ -2,3 +2,12 @@ export const DefaultAppName = 'app';
 export const TemplateExt = '.mustache';
 export const ManifestoFileName = 'manifesto.json';
 export const Directive = { Copy: 'copy', Loop: 'loop', ViewModel: 'viewModel' };
+export const AllowedConfigProperties = [
+  'name',
+  'template',
+  'dir',
+  'models',
+  'ui',
+  'api',
+  'db'
+];